change source names
