+++ applications/vessel-history/public/locales/es/translations.json
@@ -7,7 +7,7 @@
     "map": "mapa",
     "offlineAccess": "acceso sin conexión",
     "offlineAccessEmptyState": "The vessels you save for offline access will appear here.",
-    "Other": "Other",
+    "other": "Other",
     "since": "desde",
     "timeRange": "período",
     "to": "hasta",
