+++ applications/vessel-history/public/locales/fr/translations.json
@@ -7,7 +7,7 @@
     "map": "carte",
     "offlineAccess": "accès hors ligne",
     "offlineAccessEmptyState": "Les navires que vous économisez pour un accès hors ligne apparaîtront ici.",
-    "Other": "Autre",
+    "other": "Autre",
     "since": "depuis le",
     "timeRange": "intervalle de temps",
     "to": "à",
