+++ applications/vessel-history/public/locales/source/translations.json
@@ -7,7 +7,7 @@
     "map": "map",
     "offlineAccess": "offline access",
     "offlineAccessEmptyState": "The vessels you save for offline access will appear here.",
-    "Other": "Other",
+    "other": "Other",
     "since": "since",
     "timeRange": "time range",
     "to": "to",
